+++ djmoney/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.1.1"
+__version__ = "3.0.0"
 
 try:
     import django