+++ tests/test_models.py
@@ -62,8 +62,8 @@ class TestVanillaMoneyField:
     @pytest.mark.parametrize(
         "model_class, kwargs, expected",
         (
-            (ModelWithVanillaMoneyField, {"money": Money("100.0")}, Money("100.0")),
-            (ModelWithVanillaMoneyField, {"money": OldMoney("100.0")}, Money("100.0")),
+            (ModelWithVanillaMoneyField, {"money": Money("100.0", "USD")}, Money("100.0", "USD")),
+            (ModelWithVanillaMoneyField, {"money": OldMoney("100.0", "USD")}, Money("100.0", "USD")),
             (BaseModel, {}, Money(0, "USD")),
             (BaseModel, {"money": "111.2"}, Money("111.2", "USD")),
             (BaseModel, {"money": Money("123", "PLN")}, Money("123", "PLN")),
@@ -118,7 +118,6 @@ class TestVanillaMoneyField:
     @pytest.mark.parametrize(
         "model_class, other_value",
         (
-            (ModelWithVanillaMoneyField, Money("100.0")),
             (BaseModel, Money(0, "USD")),
             (ModelWithDefaultAsMoney, Money("0.01", "RUB")),
             (ModelWithDefaultAsFloat, OldMoney("12.05", "PLN")),
@@ -151,27 +150,35 @@ class TestVanillaMoneyField:
             BaseModel.objects.create(money=value)
 
     @pytest.mark.parametrize("money_class", (Money, OldMoney))
-    @pytest.mark.parametrize("field_name", ("money", "second_money"))
-    def test_save_new_value(self, field_name, money_class):
-        ModelWithVanillaMoneyField.objects.create(**{field_name: money_class("100.0")})
+    def test_save_new_value_on_field_without_default(self, money_class):
+        ModelWithVanillaMoneyField.objects.create(money=money_class("100.0", "DKK"))
 
         # Try setting the value directly
         retrieved = ModelWithVanillaMoneyField.objects.get()
-        setattr(retrieved, field_name, Money(1, "DKK"))
+        retrieved.money = Money(1, "DKK")
         retrieved.save()
         retrieved = ModelWithVanillaMoneyField.objects.get()
+        assert retrieved.money == Money(1, "DKK")
 
-        assert getattr(retrieved, field_name) == Money(1, "DKK")
+    def test_save_new_value_on_field_with_default(self):
+        ModelWithDefaultAsMoney.objects.create()
+
+        # Try setting the value directly
+        retrieved = ModelWithDefaultAsMoney.objects.get()
+        retrieved.money = Money(1, "DKK")
+        retrieved.save()
+        retrieved = ModelWithDefaultAsMoney.objects.get()
+        assert retrieved.money == Money(1, "DKK")
 
     def test_rounding(self):
-        money = Money("100.0623456781123219")
+        money = Money("100.0623456781123219", "USD")
 
         instance = ModelWithVanillaMoneyField.objects.create(money=money)
         # TODO. Should instance.money be rounded too?
 
         retrieved = ModelWithVanillaMoneyField.objects.get(pk=instance.pk)
 
-        assert retrieved.money == Money("100.06")
+        assert retrieved.money == Money("100.06", "USD")
 
     @pytest.fixture(params=[Money, OldMoney])
     def objects_setup(self, request):
@@ -240,7 +247,7 @@ class TestVanillaMoneyField:
         assert ModelWithTwoMoneyFields.objects.filter(**kwargs).count() == expected
 
     def test_exact_match(self):
-        money = Money("100.0")
+        money = Money("100.0", "USD")
 
         instance = ModelWithVanillaMoneyField.objects.create(money=money)
         retrieved = ModelWithVanillaMoneyField.objects.get(money=money)
@@ -253,9 +260,9 @@ class TestVanillaMoneyField:
         ModelIssue300.objects.filter(money__created__gt=date)
 
     def test_range_search(self):
-        money = Money("3")
+        money = Money("3", "EUR")
 
-        instance = ModelWithVanillaMoneyField.objects.create(money=Money("100.0"))
+        instance = ModelWithVanillaMoneyField.objects.create(money=Money("100.0", "EUR"))
         retrieved = ModelWithVanillaMoneyField.objects.get(money__gt=money)
 
         assert instance.pk == retrieved.pk
@@ -308,6 +315,21 @@ class TestVanillaMoneyField:
         instance = ModelWithVanillaMoneyField(money=10, money_currency=None)
         with pytest.raises(TypeError, match=r"Currency code can't be None"):
             instance.save()
+
+    def test_raises_type_error_setting_currency_to_none_on_nullable_currency_field_while_having_amount(self):
+        with pytest.raises(ValueError, match=r"Missing currency value"):
+            NullMoneyFieldModel(field=10, field_currency=None)
+
+    def test_currency_field_null_switch_not_triggered_from_default_currency(self):
+        # We want a sane default behaviour and simply declaring a `MoneyField(...)`
+        # without any default value args should create non nullable amount and currency
+        # fields
+        assert ModelWithVanillaMoneyField._meta.get_field("money").null is False
+        assert ModelWithVanillaMoneyField._meta.get_field("money_currency").null is False
+
+    def test_currency_field_nullable_when_money_field_is_nullable(self):
+        assert NullMoneyFieldModel._meta.get_field("field").null is True
+        assert NullMoneyFieldModel._meta.get_field("field_currency").null is True
 
 
 @pytest.mark.parametrize(
@@ -383,7 +405,7 @@ class TestGetOrCreate:
     @pytest.mark.parametrize(
         "model, field_name, kwargs, currency",
         (
-            (ModelWithVanillaMoneyField, "money", {"money_currency": "PLN"}, "PLN"),
+            (ModelWithDefaultAsInt, "money", {"money_currency": "PLN"}, "PLN"),
             (ModelWithVanillaMoneyField, "money", {"money": Money(0, "EUR")}, "EUR"),
             (ModelWithVanillaMoneyField, "money", {"money": OldMoney(0, "EUR")}, "EUR"),
             (ModelWithSharedCurrency, "first", {"first": 10, "second": 15, "currency": "CZK"}, "CZK"),
@@ -457,9 +479,9 @@ class TestNullableCurrency:
         assert not ModelWithNullableCurrency.objects.exists()
 
     def test_fails_with_nullable_but_no_default(self):
-        with pytest.raises(IntegrityError) as exc:
+        match = r"NOT NULL constraint failed: testapp_modelwithtwomoneyfields.amount1"
+        with pytest.raises(IntegrityError, match=match):
             ModelWithTwoMoneyFields.objects.create()
-        assert str(exc.value) == "NOT NULL constraint failed: testapp_modelwithtwomoneyfields.amount1"
 
     def test_query_not_null(self):
         money = Money(100, "EUR")
@@ -603,9 +625,11 @@ class TestExpressions:
         assert ModelWithVanillaMoneyField.objects.get(integer=1).money == Money(0, "USD")
 
     def test_create_func(self):
-        instance = ModelWithVanillaMoneyField.objects.create(money=Func(Value(-10), function="ABS"))
+        instance = ModelWithVanillaMoneyField.objects.create(
+            money=Func(Value(-10), function="ABS"), money_currency="USD"
+        )
         instance.refresh_from_db()
-        assert instance.money.amount == 10
+        assert instance.money == Money(10, "USD")
 
     @pytest.mark.parametrize(
         "value, expected", ((None, None), (10, Money(10, "USD")), (Money(10, "EUR"), Money(10, "EUR")))
@@ -617,7 +641,7 @@ class TestExpressions:
 
     def test_value_create_invalid(self):
         with pytest.raises(ValidationError):
-            ModelWithVanillaMoneyField.objects.create(money=Value("string"))
+            ModelWithVanillaMoneyField.objects.create(money=Value("string"), money_currency="DKK")
 
     def test_expressions_for_non_money_fields(self):
         instance = ModelWithVanillaMoneyField.objects.create(money=Money(1, "USD"), integer=0)
@@ -636,7 +660,7 @@ def test_find_models_related_to_money_models():
 def test_allow_expression_nodes_without_money():
     """Allow querying on expression nodes that are not Money"""
     desc = "hundred"
-    ModelWithNonMoneyField.objects.create(money=Money(100.0), desc=desc)
+    ModelWithNonMoneyField.objects.create(money=Money(100.0, "USD"), desc=desc)
     instance = ModelWithNonMoneyField.objects.filter(desc=F("desc")).get()
     assert instance.desc == desc
 
@@ -869,7 +893,7 @@ def test_mixer_blend():
     except AttributeError:
         pass  # mixer doesn't work with pypy
     else:
-        instance = mixer.blend(ModelWithTwoMoneyFields)
+        instance = mixer.blend(ModelWithTwoMoneyFields, amount1_currency="EUR", amount2_currency="USD")
         assert isinstance(instance.amount1, Money)
         assert isinstance(instance.amount2, Money)